Refactor readings
Major refactor of the readout code.  Use a struct to handle readout data, makes the code less error prone.

* Pass struct to reduce allocations.
* Fix bug in power factor calcuations.
* Update use of value data to new struct.
* Operate on one phase at a time to better correlate instantaneous readings.

+++ src/readout/database.go
@@ -12,7 +12,7 @@ import (
 	"github.com/nDenerserve/SmartPi/src/smartpi"
 )
 
-func updateSQLiteDatabase(c *smartpi.Config, data []float32) {
+func updateSQLiteDatabase(c *smartpi.Config, data []float64) {
 	t := time.Now()
 	logLine := "## SQLITE Database Update ##"
 	logLine += fmt.Sprintf(t.Format(" 2006-01-02 15:04:05 "))
