+++ src/smartpi/database.go
@@ -41,7 +41,7 @@ func CreateSQlDatabase(databasedir string, t time.Time) {
 
 }
 
-func InsertData(databasedir string, t time.Time, v []float32) {
+func InsertData(databasedir string, t time.Time, v []float64) {
 	db, err := sql.Open("sqlite3", databasedir+"/smartpi_logdata_"+t.Format("200601")+".db")
 	if err != nil {
 		log.Println(err)
@@ -68,7 +68,6 @@ func InsertData(databasedir string, t time.Time, v []float32) {
 		log.Println(err)
 	}
 
-
 	stmt, err := tx.Prepare("INSERT INTO smartpi_logdata_" + t.Format("200601") + " (date, current_1, current_2, current_3, current_4, voltage_1, voltage_2, voltage_3, power_1, power_2, power_3, cosphi_1, cosphi_2, cosphi_3, frequency_1, frequency_2, frequency_3, energy_pos_1, energy_pos_2, energy_pos_3, energy_neg_1, energy_neg_2, energy_neg_3) values (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)")
 
 	if err != nil {
@@ -90,7 +89,6 @@ func ReadChartData(databasedir string, starttime time.Time, endtime time.Time) [
 	diffmonth := Monthchange(starttime, endtime)
 
 	elapsedtime := endtime
-
 
 	for i := 0; i <= diffmonth; i++ {
 
@@ -211,7 +209,7 @@ func ReadDayData(databasedir string, starttime time.Time, endtime time.Time) []*
 			val := new(MinuteValues)
 			insert := 1
 
-			entrydate,_ := time.ParseInLocation("2006-01-02T15:04:05Z",dateentry,time.Now().Location())
+			entrydate, _ := time.ParseInLocation("2006-01-02T15:04:05Z", dateentry, time.Now().Location())
 
 			for i := 0; i < len(values); i++ {
 
@@ -247,7 +245,7 @@ func ReadDayData(databasedir string, starttime time.Time, endtime time.Time) []*
 
 			if insert == 1 {
 
-				val.Date, err = time.ParseInLocation("2006-01-02T15:04:05Z",dateentry,time.Now().Location())
+				val.Date, err = time.ParseInLocation("2006-01-02T15:04:05Z", dateentry, time.Now().Location())
 				val.Current_1 = current_1
 				val.Current_2 = current_2
 				val.Current_3 = current_3
